changing organization of web app resources
