+++ saml_reader/web/app.py
@@ -1,5 +1,6 @@
 from dash_extensions.enrich import DashProxy, MultiplexerTransform
 
+
 app = DashProxy(__name__,
     external_stylesheets=[
         "https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
