+++ saml_reader/web/callbacks/__init__.py
@@ -1,0 +1,3 @@
+import saml_reader.web.callbacks.index
+import saml_reader.web.callbacks.home
+import saml_reader.web.callbacks.analyze
