+++ saml_reader/web/callbacks/home.py
@@ -1,0 +1,6 @@
+from dash.dependencies import Input, Output, State
+from dash.exceptions import PreventUpdate
+
+from saml_reader.web.app import app
+
+# There are no callbacks for this page right now
