+++ saml_reader/web/layouts/__init__.py
@@ -1,0 +1,3 @@
+import saml_reader.web.layouts.index
+import saml_reader.web.layouts.home
+import saml_reader.web.layouts.analyze
